Move actions from partial classes into controllers
