+++ src/FrontendAccountCreation.Web/Controllers/ReprocessorExporter/ApprovedPersonController.cs
@@ -12,48 +12,48 @@ using System.Collections.Generic;
 
 namespace FrontendAccountCreation.Web.Controllers.ReprocessorExporter
 {
-	[Feature(FeatureFlags.AddOrganisationCompanyHouseDirectorJourney)]
-	[Route("re-ex/organisation")]
+    [Feature(FeatureFlags.AddOrganisationCompanyHouseDirectorJourney)]
+    [Route("re-ex/organisation")]
     public partial class ApprovedPersonController : ControllerBase<OrganisationSession>
-	{
-		private readonly ISessionManager<OrganisationSession> _sessionManager;
-		private readonly ExternalUrlsOptions _urlOptions;
+    {
+        private readonly ISessionManager<OrganisationSession> _sessionManager;
+        private readonly ExternalUrlsOptions _urlOptions;
 
         public ApprovedPersonController(
             ISessionManager<OrganisationSession> sessionManager,
             IOptions<ExternalUrlsOptions> urlOptions) : base(sessionManager)
-		{
-			_sessionManager = sessionManager;
-			_urlOptions = urlOptions.Value;
-		}
+        {
+            _sessionManager = sessionManager;
+            _urlOptions = urlOptions.Value;
+        }
 
-		[HttpGet]
-		[Route(PagePath.AddAnApprovedPerson)]
-		[OrganisationJourneyAccess(PagePath.AddAnApprovedPerson)]
-		public async Task<IActionResult> AddApprovedPerson()
-		{
-			var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
-			SetBackLink(session, PagePath.AddAnApprovedPerson);
-			await _sessionManager.SaveSessionAsync(HttpContext.Session, session);
+        [HttpGet]
+        [Route(PagePath.AddAnApprovedPerson)]
+        [OrganisationJourneyAccess(PagePath.AddAnApprovedPerson)]
+        public async Task<IActionResult> AddApprovedPerson()
+        {
+            var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
+            SetBackLink(session, PagePath.AddAnApprovedPerson);
+            await _sessionManager.SaveSessionAsync(HttpContext.Session, session);
 
             var model = new AddApprovedPersonViewModel
             {
                 IsOrganisationAPartnership = session.IsOrganisationAPartnership,
                 IsInEligibleToBeApprovedPerson =
                     session.ReExCompaniesHouseSession?.IsInEligibleToBeApprovedPerson ?? false,
-				IsLimitedPartnership = session.ReExCompaniesHouseSession?.Partnership?.IsLimitedPartnership ?? false,
-				IsLimitedLiablePartnership = session.ReExCompaniesHouseSession?.Partnership?.IsLimitedLiabilityPartnership ?? false
+                IsLimitedPartnership = session.ReExCompaniesHouseSession?.Partnership?.IsLimitedPartnership ?? false,
+                IsLimitedLiablePartnership = session.ReExCompaniesHouseSession?.Partnership?.IsLimitedLiabilityPartnership ?? false
             };
 
-            return View(model); 
-		}
+            return View(model);
+        }
 
-		[HttpPost]
-		[Route(PagePath.AddAnApprovedPerson)]
-		[OrganisationJourneyAccess(PagePath.AddAnApprovedPerson)]
-		public async Task<IActionResult> AddApprovedPerson(AddApprovedPersonViewModel model)
-		{
-			var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
+        [HttpPost]
+        [Route(PagePath.AddAnApprovedPerson)]
+        [OrganisationJourneyAccess(PagePath.AddAnApprovedPerson)]
+        public async Task<IActionResult> AddApprovedPerson(AddApprovedPersonViewModel model)
+        {
+            var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
 
             if (!ModelState.IsValid)
             {
@@ -66,19 +66,19 @@ namespace FrontendAccountCreation.Web.Controllers.ReprocessorExporter
             }
 
             if (model.InviteUserOption == InviteUserOptions.BeAnApprovedPerson.ToString())
-			{
+            {
                 session.IsApprovedUser = true;
-				return await SaveSessionAndRedirect(session, nameof(YouAreApprovedPerson), PagePath.AddAnApprovedPerson, PagePath.YouAreApprovedPerson);
-			}
+                return await SaveSessionAndRedirect(session, nameof(YouAreApprovedPerson), PagePath.AddAnApprovedPerson, PagePath.YouAreApprovedPerson);
+            }
 
-			if (model.InviteUserOption == InviteUserOptions.InviteAnotherPerson.ToString())
-			{
-				return await SaveSessionAndRedirect(session, nameof(TeamMemberRoleInOrganisation),
-					PagePath.AddAnApprovedPerson, PagePath.TeamMemberRoleInOrganisation);
-			}
+            if (model.InviteUserOption == InviteUserOptions.InviteAnotherPerson.ToString())
+            {
+                return await SaveSessionAndRedirect(session, nameof(TeamMemberRoleInOrganisation),
+                    PagePath.AddAnApprovedPerson, PagePath.TeamMemberRoleInOrganisation);
+            }
 
-			return await SaveSessionAndRedirect(session, nameof(CheckYourDetails), PagePath.AddAnApprovedPerson, PagePath.CheckYourDetails);
-		}
+            return await SaveSessionAndRedirect(session, nameof(CheckYourDetails), PagePath.AddAnApprovedPerson, PagePath.CheckYourDetails);
+        }
 
         [HttpGet]
         [Route(PagePath.TeamMemberRoleInOrganisation)]
@@ -88,7 +88,7 @@ namespace FrontendAccountCreation.Web.Controllers.ReprocessorExporter
             var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
             SetBackLink(session, PagePath.TeamMemberRoleInOrganisation);
 
-			await _sessionManager.SaveSessionAsync(HttpContext.Session, session);
+            await _sessionManager.SaveSessionAsync(HttpContext.Session, session);
 
             var viewModel = new TeamMemberRoleInOrganisationViewModel();
 
@@ -106,24 +106,24 @@ namespace FrontendAccountCreation.Web.Controllers.ReprocessorExporter
                 }
             }
 
-			return session.IsOrganisationAPartnership == true
-				? View("ApprovedPersonPartnershipRole", viewModel)
-				: View(viewModel);
-		}
+            return session.IsOrganisationAPartnership == true
+                ? View("ApprovedPersonPartnershipRole", viewModel)
+                : View(viewModel);
+        }
 
-		[HttpPost]
-		[Route(PagePath.TeamMemberRoleInOrganisation)]
-		[OrganisationJourneyAccess(PagePath.TeamMemberRoleInOrganisation)]
-		public async Task<IActionResult> TeamMemberRoleInOrganisation(TeamMemberRoleInOrganisationViewModel model)
-		{
-			var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
-			if (!ModelState.IsValid)
-			{
+        [HttpPost]
+        [Route(PagePath.TeamMemberRoleInOrganisation)]
+        [OrganisationJourneyAccess(PagePath.TeamMemberRoleInOrganisation)]
+        public async Task<IActionResult> TeamMemberRoleInOrganisation(TeamMemberRoleInOrganisationViewModel model)
+        {
+            var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
+            if (!ModelState.IsValid)
+            {
                 SetBackLink(session, PagePath.TeamMemberRoleInOrganisation);
                 return session.IsOrganisationAPartnership == true
-					? View("ApprovedPersonPartnershipRole", model)
-					: View(model);
-			}
+                    ? View("ApprovedPersonPartnershipRole", model)
+                    : View(model);
+            }
 
             var companiesHouseSession = session.ReExCompaniesHouseSession ?? new();
             var members = companiesHouseSession.TeamMembers ?? new();
@@ -136,13 +136,13 @@ namespace FrontendAccountCreation.Web.Controllers.ReprocessorExporter
             }
             var queryStringId = model?.Id ?? Guid.NewGuid();
 
-			if (model.RoleInOrganisation == ReExTeamMemberRole.None)
-			{
-				if (isExistingMember)
-				{
-					members.RemoveAt(index);
-					companiesHouseSession.TeamMembers = members;
-					session.ReExCompaniesHouseSession = companiesHouseSession;
+            if (model.RoleInOrganisation == ReExTeamMemberRole.None)
+            {
+                if (isExistingMember)
+                {
+                    members.RemoveAt(index);
+                    companiesHouseSession.TeamMembers = members;
+                    session.ReExCompaniesHouseSession = companiesHouseSession;
 
                     SetFocusId(queryStringId);
                     return await SaveSessionAndRedirect(
@@ -152,28 +152,28 @@ namespace FrontendAccountCreation.Web.Controllers.ReprocessorExporter
                         PagePath.CheckYourDetails);
                 }
 
-				return await SaveSessionAndRedirect(
-					session,
-					nameof(PersonCanNotBeInvited),
-					$"{PagePath.TeamMemberRoleInOrganisation}?id={queryStringId}",
-					PagePath.ApprovedPersonPartnershipCanNotBeInvited);
-			}
+                return await SaveSessionAndRedirect(
+                    session,
+                    nameof(PersonCanNotBeInvited),
+                    $"{PagePath.TeamMemberRoleInOrganisation}?id={queryStringId}",
+                    PagePath.ApprovedPersonPartnershipCanNotBeInvited);
+            }
 
-			if (isExistingMember)
-			{
-				members[index].Role = model.RoleInOrganisation;
-			}
-			else
-			{
-				members.Add(new ReExCompanyTeamMember
-				{
-					Id = queryStringId,
-					Role = model.RoleInOrganisation
-				});
-			}
+            if (isExistingMember)
+            {
+                members[index].Role = model.RoleInOrganisation;
+            }
+            else
+            {
+                members.Add(new ReExCompanyTeamMember
+                {
+                    Id = queryStringId,
+                    Role = model.RoleInOrganisation
+                });
+            }
 
-			companiesHouseSession.TeamMembers = members;
-			session.ReExCompaniesHouseSession = companiesHouseSession;
+            companiesHouseSession.TeamMembers = members;
+            session.ReExCompaniesHouseSession = companiesHouseSession;
 
             if (isExistingMember)
             {
@@ -193,13 +193,38 @@ namespace FrontendAccountCreation.Web.Controllers.ReprocessorExporter
         }
 
         [HttpGet]
+        [Route(PagePath.TeamMemberRoleInOrganisationAdd)]
+        [OrganisationJourneyAccess(PagePath.TeamMemberRoleInOrganisation)]
+        public async Task<IActionResult> TeamMemberRoleInOrganisationAdd()
+        {
+            DeleteFocusId();
+
+            OrganisationSession? session = await _sessionManager.GetSessionAsync(HttpContext.Session);
+
+            return await SaveSessionAndRedirect(session, nameof(ApprovedPersonController.TeamMemberRoleInOrganisation),
+                PagePath.TeamMemberRoleInOrganisation, null);
+        }
+
+        [HttpGet]
+        [Route(PagePath.TeamMemberRoleInOrganisationEdit)]
+        public async Task<IActionResult> TeamMemberRoleInOrganisationEdit([FromQuery] Guid id)
+        {
+            SetFocusId(id);
+
+            OrganisationSession? session = await _sessionManager.GetSessionAsync(HttpContext.Session);
+
+            return await SaveSessionAndRedirect(session, nameof(ApprovedPersonController.TeamMemberRoleInOrganisation),
+                PagePath.TeamMemberRoleInOrganisation, null);
+        }
+
+        [HttpGet]
         [Route(PagePath.TeamMemberDetails)]
         [OrganisationJourneyAccess(PagePath.TeamMemberDetails)]
         public async Task<IActionResult> TeamMemberDetails()
         {
             Guid? id = GetFocusId();
-			if (id.HasValue)
-			{
+            if (id.HasValue)
+            {
                 SetFocusId(id.Value);
             }
             else
@@ -210,9 +235,9 @@ namespace FrontendAccountCreation.Web.Controllers.ReprocessorExporter
             var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
             SetBackLink(session, PagePath.TeamMemberDetails);
 
-			await _sessionManager.SaveSessionAsync(HttpContext.Session, session);
+            await _sessionManager.SaveSessionAsync(HttpContext.Session, session);
 
-			var index = session.ReExCompaniesHouseSession?.TeamMembers?.FindIndex(0, x => x.Id.Equals(id));
+            var index = session.ReExCompaniesHouseSession?.TeamMembers?.FindIndex(0, x => x.Id.Equals(id));
 
             if (index is >= 0)
             {
@@ -225,39 +250,51 @@ namespace FrontendAccountCreation.Web.Controllers.ReprocessorExporter
                     Email = session.ReExCompaniesHouseSession.TeamMembers[index.Value].Email
                 };
 
-				return View(viewModel);
-			}
+                return View(viewModel);
+            }
 
-			return View();
-		}
+            return View();
+        }
 
-		[HttpPost]
-		[Route(PagePath.TeamMemberDetails)]
-		[OrganisationJourneyAccess(PagePath.TeamMemberDetails)]
-		public async Task<IActionResult> TeamMemberDetails(TeamMemberViewModel model)
-		{
-			var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
-			if (!ModelState.IsValid)
-			{
+        [HttpPost]
+        [Route(PagePath.TeamMemberDetails)]
+        [OrganisationJourneyAccess(PagePath.TeamMemberDetails)]
+        public async Task<IActionResult> TeamMemberDetails(TeamMemberViewModel model)
+        {
+            var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
+            if (!ModelState.IsValid)
+            {
                 SetBackLink(session, PagePath.TeamMemberDetails);
                 return View(model);
-			}
+            }
 
-			var index = session.ReExCompaniesHouseSession.TeamMembers?.FindIndex(0, x => x.Id.Equals(model.Id));
+            var index = session.ReExCompaniesHouseSession.TeamMembers?.FindIndex(0, x => x.Id.Equals(model.Id));
 
-			if (index is >= 0)
-			{
-				// found existing team member
-				session.ReExCompaniesHouseSession.TeamMembers[index.Value].FirstName = model.FirstName;
-				session.ReExCompaniesHouseSession.TeamMembers[index.Value].LastName = model.LastName;
-				session.ReExCompaniesHouseSession.TeamMembers[index.Value].TelephoneNumber = model.Telephone;
-				session.ReExCompaniesHouseSession.TeamMembers[index.Value].Email = model.Email;
-			}
+            if (index is >= 0)
+            {
+                // found existing team member
+                session.ReExCompaniesHouseSession.TeamMembers[index.Value].FirstName = model.FirstName;
+                session.ReExCompaniesHouseSession.TeamMembers[index.Value].LastName = model.LastName;
+                session.ReExCompaniesHouseSession.TeamMembers[index.Value].TelephoneNumber = model.Telephone;
+                session.ReExCompaniesHouseSession.TeamMembers[index.Value].Email = model.Email;
+            }
 
-			// go to check invitation details summary page for all team members
-			return await SaveSessionAndRedirect(session, nameof(TeamMembersCheckInvitationDetails), PagePath.TeamMemberDetails,
-				PagePath.TeamMembersCheckInvitationDetails);
-		}
+            // go to check invitation details summary page for all team members
+            return await SaveSessionAndRedirect(session, nameof(TeamMembersCheckInvitationDetails), PagePath.TeamMemberDetails,
+                PagePath.TeamMembersCheckInvitationDetails);
+        }
+
+        [HttpGet]
+        [Route(PagePath.TeamMemberDetailsEdit)]
+        public async Task<IActionResult> TeamMemberDetailsEdit([FromQuery] Guid id)
+        {
+            SetFocusId(id);
+
+            OrganisationSession? session = await _sessionManager.GetSessionAsync(HttpContext.Session);
+
+            return await SaveSessionAndRedirect(session, nameof(ApprovedPersonController.TeamMemberDetails),
+                PagePath.TeamMemberDetails, null);
+        }
 
         /// <summary>
         /// Show team member details enetered so far
@@ -272,10 +309,10 @@ namespace FrontendAccountCreation.Web.Controllers.ReprocessorExporter
             var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
             SetBackLink(session, PagePath.TeamMembersCheckInvitationDetails);
 
-			await _sessionManager.SaveSessionAsync(HttpContext.Session, session);
+            await _sessionManager.SaveSessionAsync(HttpContext.Session, session);
 
-			return View(session.ReExCompaniesHouseSession?.TeamMembers?.Where(x => !string.IsNullOrWhiteSpace(x.FirstName)).ToList());
-		}
+            return View(session.ReExCompaniesHouseSession?.TeamMembers?.Where(x => !string.IsNullOrWhiteSpace(x.FirstName)).ToList());
+        }
 
         [HttpPost]
         [Route(PagePath.TeamMembersCheckInvitationDetails)]
@@ -286,113 +323,137 @@ namespace FrontendAccountCreation.Web.Controllers.ReprocessorExporter
             return await SaveSessionAndRedirect(session, nameof(CheckYourDetails), PagePath.TeamMembersCheckInvitationDetails, PagePath.CheckYourDetails);
         }
 
-		[HttpGet]
-		[Route(PagePath.YouAreApprovedPerson)]
-		[OrganisationJourneyAccess(PagePath.YouAreApprovedPerson)]
-		public async Task<IActionResult> YouAreApprovedPerson()
-		{
-			var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
-			SetBackLink(session, PagePath.YouAreApprovedPerson);
-			await _sessionManager.SaveSessionAsync(HttpContext.Session, session);
+        [HttpGet]
+        [Route(PagePath.TeamMembersCheckInvitationDetailsDelete)]
+        public async Task<IActionResult> TeamMembersCheckInvitationDetailsDelete([FromQuery] Guid id)
+        {
+            DeleteFocusId();
 
-			bool isPartnership = session.IsOrganisationAPartnership ?? false;
+            OrganisationSession? session = await _sessionManager.GetSessionAsync(HttpContext.Session);
+            session?.ReExCompaniesHouseSession?.TeamMembers?.RemoveAll(x => x.Id == id);
 
-			var approvedPersonViewModel = new ApprovedPersonViewModel
-			{
-				IsLimitedLiabilityPartnership = isPartnership && (session.ReExCompaniesHouseSession?.Partnership?.IsLimitedLiabilityPartnership ?? false),
-				IsLimitedPartnership = isPartnership && (session.ReExCompaniesHouseSession?.Partnership?.IsLimitedPartnership ?? false)
-			};
+            return await SaveSessionAndRedirect(session, nameof(TeamMembersCheckInvitationDetails),
+                PagePath.TeamMembersCheckInvitationDetails, null);
+        }
 
-			return View(approvedPersonViewModel);
-		}
+        [HttpGet]
+        [Route(PagePath.YouAreApprovedPerson)]
+        [OrganisationJourneyAccess(PagePath.YouAreApprovedPerson)]
+        public async Task<IActionResult> YouAreApprovedPerson()
+        {
+            var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
+            SetBackLink(session, PagePath.YouAreApprovedPerson);
+            await _sessionManager.SaveSessionAsync(HttpContext.Session, session);
 
-		[HttpPost]
-		[Route(PagePath.YouAreApprovedPerson)]
-		[OrganisationJourneyAccess(PagePath.YouAreApprovedPerson)]
-		public async Task<IActionResult> YouAreApprovedPerson(bool inviteApprovedPerson)
-		{
-			var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
-			SetBackLink(session, PagePath.YouAreApprovedPerson);
+            bool isPartnership = session.IsOrganisationAPartnership ?? false;
 
-			var nextPage = inviteApprovedPerson
-				? PagePath.TeamMemberRoleInOrganisation
-				: PagePath.CheckYourDetails;
+            var approvedPersonViewModel = new ApprovedPersonViewModel
+            {
+                IsLimitedLiabilityPartnership = isPartnership && (session.ReExCompaniesHouseSession?.Partnership?.IsLimitedLiabilityPartnership ?? false),
+                IsLimitedPartnership = isPartnership && (session.ReExCompaniesHouseSession?.Partnership?.IsLimitedPartnership ?? false)
+            };
 
-			var nextAction = inviteApprovedPerson
-				? nameof(TeamMemberRoleInOrganisation)
-				: nameof(CheckYourDetails);
+            return View(approvedPersonViewModel);
+        }
 
-			return await SaveSessionAndRedirect(session, nextAction, PagePath.YouAreApprovedPerson, nextPage);
-		}
+        [HttpPost]
+        [Route(PagePath.YouAreApprovedPerson)]
+        [OrganisationJourneyAccess(PagePath.YouAreApprovedPerson)]
+        public async Task<IActionResult> YouAreApprovedPerson(bool inviteApprovedPerson)
+        {
+            var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
+            SetBackLink(session, PagePath.YouAreApprovedPerson);
 
-		[HttpGet]
-		[Route(PagePath.MemberPartnership)]
-		[OrganisationJourneyAccess(PagePath.MemberPartnership)]
-		public async Task<IActionResult> MemberPartnership()
-		{
-			var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
-			SetBackLink(session, PagePath.MemberPartnership);
-			await _sessionManager.SaveSessionAsync(HttpContext.Session, session);
+            var nextPage = inviteApprovedPerson
+                ? PagePath.TeamMemberRoleInOrganisation
+                : PagePath.CheckYourDetails;
 
-			var viewModel = new IsMemberPartnershipViewModel();
-			return View(viewModel);
-		}
+            var nextAction = inviteApprovedPerson
+                ? nameof(TeamMemberRoleInOrganisation)
+                : nameof(CheckYourDetails);
 
-		[HttpPost]
-		[Route(PagePath.MemberPartnership)]
-		[OrganisationJourneyAccess(PagePath.MemberPartnership)]
-		public async Task<IActionResult> MemberPartnership(IsMemberPartnershipViewModel model)
-		{
-			var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
-			if (!ModelState.IsValid)
-			{
-				SetBackLink(session, PagePath.MemberPartnership);
-				return View(model);
-			}
+            return await SaveSessionAndRedirect(session, nextAction, PagePath.YouAreApprovedPerson, nextPage);
+        }
 
-			if (model.IsMemberPartnership == YesNoAnswer.Yes)
-			{
-				return await SaveSessionAndRedirect(session, "PartnerDetails", PagePath.MemberPartnership, PagePath.PartnerDetails);
-			}
+        [HttpGet]
+        [Route(PagePath.MemberPartnership)]
+        [OrganisationJourneyAccess(PagePath.MemberPartnership)]
+        public async Task<IActionResult> MemberPartnership()
+        {
+            var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
+            SetBackLink(session, PagePath.MemberPartnership);
+            await _sessionManager.SaveSessionAsync(HttpContext.Session, session);
 
-			return await SaveSessionAndRedirect(session, "CanNotInviteThisPerson", PagePath.MemberPartnership, PagePath.CanNotInviteThisPerson);
-		}
+            var viewModel = new IsMemberPartnershipViewModel();
+            return View(viewModel);
+        }
 
-		[HttpGet]
-		[Route(PagePath.CheckYourDetails)]
+        [HttpPost]
+        [Route(PagePath.MemberPartnership)]
+        [OrganisationJourneyAccess(PagePath.MemberPartnership)]
+        public async Task<IActionResult> MemberPartnership(IsMemberPartnershipViewModel model)
+        {
+            var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
+            if (!ModelState.IsValid)
+            {
+                SetBackLink(session, PagePath.MemberPartnership);
+                return View(model);
+            }
+
+            if (model.IsMemberPartnership == YesNoAnswer.Yes)
+            {
+                return await SaveSessionAndRedirect(session, "PartnerDetails", PagePath.MemberPartnership, PagePath.PartnerDetails);
+            }
+
+            return await SaveSessionAndRedirect(session, "CanNotInviteThisPerson", PagePath.MemberPartnership, PagePath.CanNotInviteThisPerson);
+        }
+
+        [HttpGet]
+        [Route(PagePath.MemberPartnershipAdd)]
+        public async Task<IActionResult> MemberPartnershipAdd()
+        {
+            DeleteFocusId();
+
+            OrganisationSession? session = await _sessionManager.GetSessionAsync(HttpContext.Session);
+
+            return await SaveSessionAndRedirect(session, nameof(MemberPartnership),
+                PagePath.YouAreApprovedPerson, PagePath.MemberPartnership);
+        }
+
+        [HttpGet]
+        [Route(PagePath.CheckYourDetails)]
         [OrganisationJourneyAccess(PagePath.CheckYourDetails)]
-		public async Task<IActionResult> CheckYourDetails()
-		{
-			var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
-			ViewBag.MakeChangesToYourLimitedCompanyLink = _urlOptions.MakeChangesToYourLimitedCompany;
+        public async Task<IActionResult> CheckYourDetails()
+        {
+            var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
+            ViewBag.MakeChangesToYourLimitedCompanyLink = _urlOptions.MakeChangesToYourLimitedCompany;
 
-			var viewModel = new ReExCheckYourDetailsViewModel
-			{
-				IsCompaniesHouseFlow = session.IsCompaniesHouseFlow,
-				IsRegisteredAsCharity = session.IsTheOrganisationCharity,
-				OrganisationType = session.OrganisationType,
-				IsTradingNameDifferent = session.IsTradingNameDifferent,
-				Nation = session.UkNation
-			};
-			if (viewModel.IsCompaniesHouseFlow)
-			{
-				viewModel.BusinessAddress = session.ReExCompaniesHouseSession.Company.BusinessAddress;
-				viewModel.CompanyName = session.ReExCompaniesHouseSession?.Company.Name;
-				viewModel.CompaniesHouseNumber = session.ReExCompaniesHouseSession?.Company.CompaniesHouseNumber;
-				viewModel.RoleInOrganisation = session.ReExCompaniesHouseSession?.RoleInOrganisation;
+            var viewModel = new ReExCheckYourDetailsViewModel
+            {
+                IsCompaniesHouseFlow = session.IsCompaniesHouseFlow,
+                IsRegisteredAsCharity = session.IsTheOrganisationCharity,
+                OrganisationType = session.OrganisationType,
+                IsTradingNameDifferent = session.IsTradingNameDifferent,
+                Nation = session.UkNation
+            };
+            if (viewModel.IsCompaniesHouseFlow)
+            {
+                viewModel.BusinessAddress = session.ReExCompaniesHouseSession.Company.BusinessAddress;
+                viewModel.CompanyName = session.ReExCompaniesHouseSession?.Company.Name;
+                viewModel.CompaniesHouseNumber = session.ReExCompaniesHouseSession?.Company.CompaniesHouseNumber;
+                viewModel.RoleInOrganisation = session.ReExCompaniesHouseSession?.RoleInOrganisation;
                 viewModel.IsOrganisationAPartnership = session.IsOrganisationAPartnership ?? false;
                 viewModel.LimitedPartnershipPartners = session.ReExCompaniesHouseSession?.Partnership?.LimitedPartnership?.Partners;
                 viewModel.IsLimitedLiabilityPartnership = session.ReExCompaniesHouseSession?.Partnership?.IsLimitedLiabilityPartnership ?? false;
-
             }
-			if (session.ReExManualInputSession != null)
-			{
-				viewModel.TradingName = session.ReExManualInputSession.TradingName;
-			}
-			viewModel.reExCompanyTeamMembers = session.ReExCompaniesHouseSession?.TeamMembers;
-			_sessionManager.SaveSessionAsync(HttpContext.Session, session);
+            if (session.ReExManualInputSession != null)
+            {
+                viewModel.TradingName = session.ReExManualInputSession.TradingName;
+            }
+            viewModel.reExCompanyTeamMembers = session.ReExCompaniesHouseSession?.TeamMembers;
+            _sessionManager.SaveSessionAsync(HttpContext.Session, session);
 
-			return View(viewModel);
+            return View(viewModel);
         }
 
         [HttpPost]
@@ -402,26 +463,26 @@ namespace FrontendAccountCreation.Web.Controllers.ReprocessorExporter
         {
             var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
             return await SaveSessionAndRedirect(session, nameof(OrganisationController.Declaration), PagePath.CheckYourDetails, PagePath.Declaration, nameof(OrganisationController).Replace("Controller", ""));
-		}
+        }
 
-		[HttpGet]
-		[Route(PagePath.ApprovedPersonPartnershipCanNotBeInvited)]
-		[OrganisationJourneyAccess(PagePath.ApprovedPersonPartnershipCanNotBeInvited)]
-		public async Task<IActionResult> PersonCanNotBeInvited([FromQuery] Guid id)
-		{
-			var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
-			SetBackLink(session, PagePath.ApprovedPersonPartnershipCanNotBeInvited);
-			await _sessionManager.SaveSessionAsync(HttpContext.Session, session);
+        [HttpGet]
+        [Route(PagePath.ApprovedPersonPartnershipCanNotBeInvited)]
+        [OrganisationJourneyAccess(PagePath.ApprovedPersonPartnershipCanNotBeInvited)]
+        public async Task<IActionResult> PersonCanNotBeInvited([FromQuery] Guid id)
+        {
+            var session = await _sessionManager.GetSessionAsync(HttpContext.Session);
+            SetBackLink(session, PagePath.ApprovedPersonPartnershipCanNotBeInvited);
+            await _sessionManager.SaveSessionAsync(HttpContext.Session, session);
 
-			return View(new LimitedPartnershipPersonCanNotBeInvitedViewModel { Id = id });
-		}
+            return View(new LimitedPartnershipPersonCanNotBeInvitedViewModel { Id = id });
+        }
 
-		[HttpPost]
-		[Route(PagePath.ApprovedPersonPartnershipCanNotBeInvited)]
-		[OrganisationJourneyAccess(PagePath.ApprovedPersonPartnershipCanNotBeInvited)]
-		public IActionResult PersonCanNotBeInvited(LimitedPartnershipPersonCanNotBeInvitedViewModel model)
-		{
-			return RedirectToAction("CheckYourDetails", "AccountCreation");
-		}
-	}
+        [HttpPost]
+        [Route(PagePath.ApprovedPersonPartnershipCanNotBeInvited)]
+        [OrganisationJourneyAccess(PagePath.ApprovedPersonPartnershipCanNotBeInvited)]
+        public IActionResult PersonCanNotBeInvited(LimitedPartnershipPersonCanNotBeInvitedViewModel model)
+        {
+            return RedirectToAction("CheckYourDetails", "AccountCreation");
+        }
+    }
 }
